update jsbeautifier to 1.7.4
