+++ libs/jsbeautifier/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '1.6.2'
+__version__ = '1.7.4'
